[*] Stapisdk: Force rebuild stapisdk apilib

+++ src/script/002_rebuild_apilib.sh
@@ -1,3 +1,4 @@
 #!/bin/sh
 
-rm -f $STSDKROOT/.makeapilib
+#up
+[ "$STSDKROOT" ] && rm -f $STSDKROOT/.makeapilib
